Improve CLI autopilot
