+++ package.json
@@ -1,44 +1,46 @@
 {
-	"name": "web-automation-agent",
-	"version": "0.0.1",
-	"type": "module",
-	"description": "This is an example of an Apify actor.",
-	"dependencies": {
-		"@crawlee/core": "^3.5.4",
-		"apify": "^3.0.0",
-		"cheerio": "^1.0.0-rc.12",
-		"crawlee": "^3.0.0",
-		"express": "^4.18.2",
-		"gpt-3-encoder": "^1.1.4",
-		"langchain": "^0.0.147",
-		"langsmith": "^0.0.41",
-		"puppeteer": "*",
-		"puppeteer-screen-recorder": "^2.1.2",
-		"zod": "^3.22.2"
-	},
-	"devDependencies": {
-		"@apify/eslint-config-ts": "^0.2.3",
-		"@apify/tsconfig": "^0.1.0",
-		"@jest/globals": "^29.5.0",
-		"@types/express": "^4.17.17",
-		"@types/jest": "^29.5.0",
-		"@typescript-eslint/eslint-plugin": "^5.32.0",
-		"@typescript-eslint/parser": "^5.32.0",
-		"eslint": "^8.20.0",
-		"jest": "^29.5.0",
-		"ts-jest": "^29.1.0",
-		"ts-node": "^10.9.1",
-		"typescript": "^4.7.4"
-	},
-	"scripts": {
-		"start": "npm run start:dev",
-		"start:prod": "node dist/main.js",
-		"start:dev": "ts-node-esm -T src/main.ts",
-		"build": "tsc",
-		"lint": "eslint ./src --ext .ts",
-		"lint:fix": "eslint ./src --ext .ts --fix",
-		"test": "jest"
-	},
-	"author": "It's not you it's me",
-	"license": "ISC"
+  "name": "web-automation-agent",
+  "version": "0.0.1",
+  "type": "module",
+  "description": "This is an example of an Apify actor.",
+  "dependencies": {
+    "@crawlee/core": "^3.5.4",
+    "apify": "^3.0.0",
+    "cheerio": "^1.0.0-rc.12",
+    "crawlee": "^3.0.0",
+    "express": "^4.18.2",
+    "gpt-3-encoder": "^1.1.4",
+    "langchain": "^0.0.147",
+    "langsmith": "^0.0.41",
+    "puppeteer": "*",
+    "puppeteer-screen-recorder": "^2.1.2",
+    "zod": "^3.22.2"
+  },
+  "devDependencies": {
+    "@apify/eslint-config-ts": "^0.2.3",
+    "@apify/tsconfig": "^0.1.0",
+    "@jest/globals": "^29.5.0",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
+    "eslint": "^8.20.0",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
+  },
+  "scripts": {
+    "start": "npm run start:dev",
+    "start:prod": "node dist/main.js",
+    "start:dev": "ts-node-esm -T src/main.ts",
+    "build": "tsc",
+    "lint": "eslint ./src --ext .ts",
+    "lint:fix": "eslint ./src --ext .ts --fix",
+    "test": "jest",
+    "autopilot": "npm run build && node dist/cli.js",
+    "autopilot:dev": "ts-node-esm -T src/cli.ts"
+  },
+  "author": "It's not you it's me",
+  "license": "ISC"
 }
